separate multi GPU manager class
