+++ otx/cli/tools/train.py
@@ -16,21 +16,10 @@
 
 import argparse
 import os
-import signal
-import sys
-import threading
-import time
-from typing import List, Optional
 
-import torch
-import torch.distributed as dist
-import torch.multiprocessing as mp
-
-from otx.api.configuration import ConfigurableParameters
 from otx.api.configuration.helper import create
 from otx.api.entities.inference_parameters import InferenceParameters
 from otx.api.entities.model import ModelEntity
-from otx.api.entities.model_template import ModelTemplate, TaskType
 from otx.api.entities.resultset import ResultSetEntity
 from otx.api.entities.subset import Subset
 from otx.api.entities.task_environment import TaskEnvironment
@@ -48,6 +37,7 @@ from otx.cli.utils.io import (
     read_label_schema,
     save_model_data,
 )
+from otx.cli.utils.multi_gpu import MultiGPUManager
 from otx.cli.utils.parser import (
     add_hyper_parameters_sub_parser,
     gen_params_dict_from_args,
@@ -221,7 +211,7 @@ def main():
         task = task_class(task_environment=environment, output_path=args.save_logs_to)
 
     if args.gpus:
-        multigpu_manager = MultiGPUManager(args.gpus, str(args.multi_gpu_port))
+        multigpu_manager = MultiGPUManager(main, args.gpus, str(args.multi_gpu_port))
         if multigpu_manager.is_available(template):
             multigpu_manager.setup_multi_gpu_train(task.project_path, hyper_parameters if args.enable_hpo else None)
 
@@ -251,213 +241,6 @@ def main():
 
     if args.gpus:
         multigpu_manager.finalize()
-
-
-class MultiGPUManager:
-    """Class to manage multi-GPU training."""
-
-    def __init__(self, gpu_ids: str, multi_gpu_port: str):
-        self._gpu_ids = self.get_gpu_ids(gpu_ids)
-        self._multi_gpu_port = multi_gpu_port
-        self._main_pid = os.getpid()
-        self._processes: Optional[List[mp.Process]] = None
-
-    @staticmethod
-    def get_gpu_ids(gpus: str) -> List[int]:
-        """Get proper GPU indices form `--gpu` arguments.
-
-        Given `--gpus` argument, exclude inappropriate indices and transform to list of int format.
-
-        Args:
-            gpus (str): GPU indices to use. Format should be Comma-separated indices.
-
-        Returns:
-            List[int]:
-                list including proper GPU indices.
-        """
-        num_available_gpu = torch.cuda.device_count()
-        gpu_ids = []
-        for gpu_id in gpus.split(","):
-            if not gpu_id.isnumeric():
-                raise RuntimeError("--gpus argument should be numbers separated by ','.")
-            gpu_ids.append(int(gpu_id))
-
-        wrong_gpus = []
-        for gpu_idx in gpu_ids:
-            if gpu_idx >= num_available_gpu:
-                wrong_gpus.append(gpu_idx)
-
-        for wrong_gpu in wrong_gpus:
-            gpu_ids.remove(wrong_gpu)
-
-        if wrong_gpus:
-            print(f"Wrong gpu indeces are excluded. {','.join([str(val) for val in gpu_ids])} GPU will be used.")
-
-        return gpu_ids
-
-    def is_available(self, template: ModelTemplate) -> bool:
-        """Check multi GPU training is available.
-
-        Args:
-            template (ModelTemplate): template for training.
-
-        Returns:
-            bool:
-                whether multi GPU training is available.
-        """
-        return (
-            len(self._gpu_ids) > 1
-            and not template.task_type.is_anomaly
-            and template.task_type not in (TaskType.ACTION_CLASSIFICATION, TaskType.ACTION_DETECTION)
-        )
-
-    def setup_multi_gpu_train(
-        self, output_path: str, optimized_hyper_parameters: Optional[ConfigurableParameters] = None
-    ):
-        """Carry out what should be done to run multi GPU training.
-
-        Args:
-            output_path (str): output path where task output are saved.
-            optimized_hyper_parameters (ConfigurableParameters or None): hyper parameters reflecting HPO result.
-        """
-        if optimized_hyper_parameters is not None:  # if HPO is executed, optimized HPs are applied to child processes
-            self._set_optimized_hp_for_child_process(optimized_hyper_parameters)
-
-        self._processes = self._spawn_multi_gpu_processes(output_path)
-
-        signal.signal(signal.SIGINT, self._terminate_signal_handler)
-        signal.signal(signal.SIGTERM, self._terminate_signal_handler)
-
-        self.initialize_multigpu_train(0, self._gpu_ids, self._multi_gpu_port)
-
-        threading.Thread(target=self._check_child_processes_alive, daemon=True).start()
-
-    def finalize(self):
-        """Join all child processes."""
-        if self._processes is not None:
-            for p in self._processes:
-                p.join()
-
-    @staticmethod
-    def initialize_multigpu_train(rank: int, gpu_ids: List[int], multi_gpu_port: str):
-        """Initilization for multi GPU training.
-
-        Args:
-            rank (int): index of multi GPU processes.
-            gpu_ids (List[int]): list including which GPU indeces will be used.
-            multi_gpu_port (str): port for communication between multi GPU processes.
-        """
-        os.environ["MASTER_ADDR"] = "localhost"
-        os.environ["MASTER_PORT"] = multi_gpu_port
-        torch.cuda.set_device(gpu_ids[rank])
-        dist.init_process_group(backend="nccl", world_size=len(gpu_ids), rank=rank)
-        print(f"dist info world_size = {dist.get_world_size()}, rank = {dist.get_rank()}")
-
-    @staticmethod
-    def run_child_process(rank: int, gpu_ids: List[int], output_path: str, multi_gpu_port: str):
-        """Function for multi GPU child process to execute.
-
-        Args:
-            rank (int): index of multi GPU processes.
-            gpu_ids (List[int]): list including which GPU indeces will be used.
-            output_path (str): output path where task output are saved.
-            multi_gpu_port (str): port for communication between multi GPU processes.
-        """
-        gpus_arg_idx = sys.argv.index("--gpus")
-        for _ in range(2):
-            sys.argv.pop(gpus_arg_idx)
-        if "--enable-hpo" in sys.argv:
-            sys.argv.remove("--enable-hpo")
-        MultiGPUManager.set_arguments_to_argv("--save-logs-to", output_path)
-
-        MultiGPUManager.initialize_multigpu_train(rank, gpu_ids, multi_gpu_port)
-
-        main()
-
-    @staticmethod
-    def set_arguments_to_argv(key: str, value: Optional[str] = None, after_params: bool = False):
-        """Add arguments at proper position in `sys.argv`.
-
-        Args:
-            key (str): arguement key.
-            value (str or None): argument value.
-            after_params (bool): whether argument should be after `param` or not.
-        """
-        if key in sys.argv:
-            if value is not None:
-                sys.argv[sys.argv.index(key) + 1] = value
-        else:
-            if not after_params and "params" in sys.argv:
-                sys.argv.insert(sys.argv.index("params"), key)
-                if value is not None:
-                    sys.argv.insert(sys.argv.index("params"), value)
-            else:
-                if after_params and "params" not in sys.argv:
-                    sys.argv.append("params")
-                if value is not None:
-                    sys.argv.extend([key, value])
-                else:
-                    sys.argv.append(key)
-
-    def _spawn_multi_gpu_processes(self, output_path: str) -> List[mp.Process]:
-        processes = []
-        spawned_mp = mp.get_context("spawn")
-        for rank in range(1, len(self._gpu_ids)):
-            task_p = spawned_mp.Process(
-                target=MultiGPUManager.run_child_process, args=(rank, self._gpu_ids, output_path, self._multi_gpu_port)
-            )
-            task_p.start()
-            processes.append(task_p)
-
-        return processes
-
-    def _terminate_signal_handler(self, signum, _frame):
-        # This code prevents child processses from being killed unintentionally by proccesses forked from main process
-        if self._main_pid != os.getpid():
-            sys.exit()
-
-        self._kill_child_process()
-
-        singal_name = {2: "SIGINT", 15: "SIGTERM"}
-        print(f"{singal_name[signum]} is sent. process exited.")
-
-        sys.exit(1)
-
-    def _kill_child_process(self):
-        if self._processes is None:
-            return
-
-        for process in self._processes:
-            print(f"Kill child process {process.pid}")
-            try:
-                process.kill()
-            except Exception:  # pylint: disable=broad-except
-                pass
-
-    def _set_optimized_hp_for_child_process(self, hyper_parameters: ConfigurableParameters):
-        self.set_arguments_to_argv(
-            "--learning_parameters.learning_rate",
-            str(hyper_parameters.learning_parameters.learning_rate),  # type: ignore[attr-defined]
-            True,
-        )
-        self.set_arguments_to_argv(
-            "--learning_parameters.batch_size",
-            str(hyper_parameters.learning_parameters.batch_size),  # type: ignore[attr-defined]
-            True,
-        )
-
-    def _check_child_processes_alive(self):
-        child_is_running = True
-        while child_is_running:
-            time.sleep(1)
-            for p in self._processes:
-                if not p.is_alive() and p.exitcode != 0:
-                    child_is_running = False
-                    break
-
-        print("Some of child processes are terminated abnormally. process exits.")
-        self._kill_child_process()
-        os.kill(self._main_pid, signal.SIGKILL)
 
 
 if __name__ == "__main__":
